Remove vuex-perist and adjust jwt auth setup
vuex-persist was causing some issues saving the whole store in localstorage. removing this also revealed a bug with the post-login redirect that was causing a full page refresh and dumping the store, so now that step just does a router push which retains the store with the jwt

this change does NOT persist the jwt over a refresh. there are a few ways to do that, none of them are trivial.